backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.1.0
python-decouple==3.8
dj-rest-auth==5.0.2
django-allauth==0.54.0

